pyproject.toml
[tool.poetry]
name = "ct-hackathon-backend"
version = "0.1.0"
description = ""
authors = ["Maxim Solodilov <sldmxm@gmail.com>"]
readme = "README.md"
packages = [{include = "ct_hackathon_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "4.1"
djangorestframework = "^3.14.0"
pillow = "^10.1.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py311']
quote = 'single'
include = '\.pyi?$'
exclude = '''
/(
\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
| dist
)/
'''
